✨feat: update bill-related DTOs and entities to use Double for amount fields

+++ src/main/java/com/homeprotectors/backend/dto/bill/BillListItemResponse.java
@@ -11,5 +11,5 @@ public class BillListItemResponse {
     private Long id;
     private String name;
     private Integer dueDate; // null 가능
-    private Integer amount;  // 화면에 표시할 그 달 기준 금액
+    private Double amount;  // 화면에 표시할 그 달 기준 금액
 }
